GlobalActions: Redesign Power Menu
Give it a fresh coat of paint with blur (optional ofc), and translucency.
Moreover, we changed back to making the emergency button bright red.
Imagine this: Some old grandpa gets a heart attack and needs help asap, and instead of calling an ambulance fast, he struggles to find the emergency button and doesnt get the help he needs. Since we love accessibility in critical scenarios and old grandpas, it should be easily accessible.

Signed-off-by: NotThatBlueX <szabence@protonmail.com>
Change-Id: I485ec240c5a78afd695d45a42947af93a3cb8468